+++ cli/cmd/migrate.go
@@ -25,7 +25,10 @@ var migrateCmd = &cobra.Command{
 	Short:        "Migrate stack",
 	Long:         "Run migration tasks for stack with given name",
 	SilenceUsage: true,
-	PreRunE:      cmd.Validate(cobra.ExactArgs(1), cmd.IsStackNameDNSCharset),
+	PreRunE: cmd.Validate(
+		cobra.ExactArgs(1),
+		cmd.IsStackNameDNSCharset,
+	),
 	RunE: func(cmd *cobra.Command, args []string) error {
 		stackName := args[0]
 		return runMigrate(cmd, stackName)